# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pokerbots Competition API
  version: "1.0.0"
  description: |
    Student-facing and admin API for the Berkeley Pokerbots competition.

    * Base path: `/api/v1`
    * Auth: `X-Api-Key: <team_api_key>` **or** `Authorization: Bearer <team_api_key>`.
      Admin routes accept `X-Admin-Key: <admin_key>` or the key of a team flagged `is_admin`.
    * Errors always use the envelope `{ "error": { "code", "message", "details" } }` with codes
      `unauthorized`, `forbidden`, `not_found`, `validation_error`, `conflict`, `rate_limited`,
      `payload_too_large`, `not_ready`, `internal_error`.
    * Bot protocol: see `docs/BOT_PROTOCOL.md`.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
    Bearer:
      type: http
      scheme: bearer
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Manifest:
      type: object
      required: [entrypoint]
      properties:
        language: { type: string, default: other, description: "python | rust | cpp | java | other (informational)" }
        entrypoint: { type: string, description: "Path inside the artifact, e.g. bot.py or ./bot" }
        runtime: { type: string, default: auto, enum: [auto, native, python3, node, java] }
        args: { type: array, items: { type: string }, default: [] }
        build:
          type: array
          items: { type: string }
          default: []
          description: "Build command run in the sandbox before validation (argv). Empty = default per-runtime check (py_compile / node --check)."
        build_timeout_secs: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        protocol_version: { type: string, default: "1" }
    CheckReport:
      type: object
      description: Result of one validation stage.
      properties:
        stage: { type: string, enum: [build, smoke, trial] }
        status: { type: string, enum: [pending, running, passed, warned, failed, skipped] }
        started_at: { type: string, nullable: true }
        finished_at: { type: string, nullable: true }
        summary: { type: string }
        details: { type: object, additionalProperties: true }
    SmokeReport:
      type: object
      properties:
        passed: { type: boolean }
        first_latency_ms: { type: integer }
        latency_ms: { type: integer }
        reason: { type: string, nullable: true, enum: [timeout, invalid_json, illegal_action, crashed, setup_failed, null] }
        message: { type: string }
        stderr_tail: { type: string }
        actions: { type: array, items: { $ref: '#/components/schemas/Action' } }
    Action:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [Fold, Check, Call, BetTo, RaiseTo, AllIn] }
        amount: { type: integer, description: "Required for BetTo/RaiseTo: total committed on this street after acting" }
    Submission:
      type: object
      properties:
        submission_id: { type: string, example: sub_tm_ab12cd34ef56_0007 }
        team_id: { type: string }
        seq: { type: integer }
        status: { type: string, enum: [pending, building, smoke_testing, trial_running, validated, rejected, deleted] }
        in_progress: { type: boolean, description: still moving through the validation pipeline }
        protocol_version: { type: string }
        manifest: { $ref: '#/components/schemas/Manifest' }
        artifact_sha256: { type: string }
        artifact_size: { type: integer }
        created_at: { type: string, format: date-time }
        validated_at: { type: string, format: date-time, nullable: true }
        smoke_test: { $ref: '#/components/schemas/SmokeReport', nullable: true }
        checks: { type: array, items: { $ref: '#/components/schemas/CheckReport' } }
        auto_activate: { type: boolean }
        active: { type: boolean }
...
  responses:
...
    Forbidden:
      description: Not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
security:
  - ApiKey: []
  - Bearer: []
paths:
...
  /submissions:
    post:
      tags: [submissions]
      summary: Upload a bot artifact (multipart) and run the validation pipeline (build → smoke → trial)
      description: |
        `artifact`: a zip archive or a single file (max `max_upload_bytes`, default 50 MB).
        `manifest`: JSON string (see Manifest). Zips whose only top-level entry is a folder are
        hoisted automatically. Validation runs three stages — build/compile check, protocol smoke
        test, and a short trial match against reference bots — recorded in `checks`. On success
        the submission is `validated`; it becomes active only via
        `POST /submissions/{id}/activate` or when `activate=true` was sent with the upload.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [artifact, manifest]
              properties:
                artifact: { type: string, format: binary }
                manifest: { type: string, description: JSON-encoded Manifest }
                activate: { type: string, enum: ["true", "false"], description: activate automatically once validated }
      responses:
        "201":
          description: Validated
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        "202":
          description: Accepted but validation has not finished yet; poll GET /submissions/{id}
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        "413":
          description: Artifact too large
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        "422":
          description: Validation failed or manifest invalid. `details` carries `submission_id`, `stage` (build | smoke | trial), `stage_summary`, `checks`, `reason` (timeout | invalid_json | illegal_action | crashed | setup_failed), `message`, `stderr_tail`.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [submissions]
      summary: List own submissions (newest first)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: integer }, description: "seq cursor from next_cursor" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  submissions: { type: array, items: { $ref: '#/components/schemas/Submission' } }
                  next_cursor: { type: integer, nullable: true }
                  active_submission_id: { type: string, nullable: true }
  /submissions/{submission_id}:
    parameters: [{ name: submission_id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [submissions]
      summary: Submission metadata (own team only)
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [submissions]
      summary: Soft-delete an inactive submission
      responses:
        "204": { description: Deleted }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
  /submissions/{submission_id}/activate:
    post:
      tags: [submissions]
      summary: Make this the team's single active bot (deactivates the previous one)
      parameters: [{ name: submission_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  team_id: { type: string }
                  active_submission_id: { type: string }
                  previous_submission_id: { type: string, nullable: true }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
...